refactor sendwyre service
