+++ src/database/Database.ts
@@ -2,12 +2,13 @@
 const MongoClient = require('mongodb').MongoClient;
 const uri = "mongodb+srv://root:root@cluster0-sttce.mongodb.net/test?retryWrites=true&w=majority";
 
-export const client = new MongoClient(uri, { useNewUrlParser: true });
+const client = new MongoClient(uri, { useNewUrlParser: true });
 
-// client.connect((err: any) => {
-//     let db = client.db("myproject");
-//     let cursor = db.collection('accounts').find({userId: 292});
-//     console.log(cursor.For);
-//     // perform actions on the collection object
-//     client.close();
-// });
+export const connection =  async () => {
+    try {
+        const conn = await client.connect();
+        return conn;
+    } catch (err) {
+        console.log(err)
+    }
+}
