+++ src/modules/accounts/index.ts
@@ -1,11 +1,6 @@
 //@ts-ignore
 import express from 'express';
 import AccountsController from './AccountsController';
-// import middlewares from '../../middlewares';
-// import GuestHouseValidator from '../../middlewares/guestHouseValidator';
-
-
-// const { authenticate, RoleValidator } = middlewares;
 const Router = express.Router();
 
 Router.get(
@@ -23,20 +18,20 @@ Router.post(
   AccountsController.createPaymentMethod
 );
 
-Router.get(
-  '/paymentMethods',
-  AccountsController.getPaymentMethod
-);
+// Router.get(
+//   '/paymentMethods',
+//   AccountsController.getPaymentMethod
+// );
 
-Router.get(
-  '/allPaymentMethods',
-  AccountsController.getAllPaymentMethods
-);
+// Router.get(
+//   '/allPaymentMethods',
+//   AccountsController.getAllPaymentMethods
+// );
 
-Router.put(
-  '/accounts',
-  AccountsController.updateAccount
-);
+// Router.put(
+//   '/accounts',
+//   AccountsController.updateAccount
+// );
 
 Router.post(
   '/attachBlockChain',
